Update tls.lua: SSLv23_method() → TLS_method()
TLS_method() is the general-purpose version-flexible SSL/TLS methods. The actual protocol version used will be negotiated to the highest version mutually supported by the client and the server. The supported protocols are SSLv3, TLSv1, TLSv1.1, TLSv1.2 and TLSv1.3. Applications should use these methods, and avoid the version-specific methods, which are deprecated.

https://www.openssl.org/docs/manmaster/man3/SSL_CTX_new.html