Upgrade to latest OpenSSL
This commit replaces the outdated for of the OpenSSL project with that project itself. This will introduce loss of functionality (eg: no more SSL v2.0 support).

With the upgrade, the outdated fork of luasec doesn't build. Updating that to its original repo resolved that (but again, is likely to result in lost features).